fix: deprecated another related file to table
