+++ src/sample/table-deprecated/data.ts
